backend/requirements.txt
Flask
Flask-Cors
PyYAML
langchain
langchain-community
langchain-ollama
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
ollama 